LayoutTests/discovery/discovery-zeroconf-available-expected.txt

Add request
WebView network services request permission 2 service(s) (origin "file://")
	Services:
	 - 2 allowed
	 - 0 denied
	 - 2 online
	 - 0 offline
Tests that available event is send when a new service requested is published.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
* register default error callback
* sent network services request
* got reply: 2 service(s) on 2 services available
* services[0]: zeroconf:_test-available._tcp
* services[1]: zeroconf:_test-available._tcp
* register onavailable for service 0
* register onavailable for service 1
* service available : TestRunner1
